+++ src/core/components/calendar/calendar.js
@@ -74,7 +74,7 @@ export default {
       toolbarCloseText: 'Done',
       footer: false,
       cssClass: null,
-      routableModals: true,
+      routableModals: false,
       view: null,
       url: 'date/',
       backdrop: null,