+++ src/core/components/color-picker/color-picker.d.ts
@@ -145,7 +145,7 @@ export namespace ColorPicker {
     navbarTitleText: string;
     /** Navbar's back link text (when opened in page). (default Back) */
     navbarBackLinkText: string;
-    /** Will add opened Color Picker to router history which gives ability to close Color Picker by going back in router history and set current route to the Color Picker modal. (default true) */
+    /** Will add opened Color Picker to router history which gives ability to close Color Picker by going back in router history and set current route to the Color Picker modal. (default false) */
     routableModals?: boolean;
     /** Color Picker modal URL that will be set as a current route. (default color/) */
     url?: string;