+++ src/core/components/picker/picker.js
@@ -49,7 +49,7 @@ export default {
       toolbar: true,
       toolbarCloseText: 'Done',
       cssClass: null,
-      routableModals: true,
+      routableModals: false,
       view: null,
       url: 'select/',
       // Render functions